feat: :sparkles: add `str_insert_at_idx` function (helpful for line editing)
